+++ Synthese/octave.cpp
@@ -11,17 +11,17 @@ Octave::Octave(const Octave &o):freq(o.freq), ampli(o.ampli), minSeuil(o.minSeui
 
 }
 
-Octave::Octave(float f, float a, float ms, float as, float fs, FoncteurWarp wPoint, FoncteurWarp wSeuil, Noise nPoint, Noise nSeuil):freq(f), ampli(a), minSeuil(ms), ampliS(as), freqS(fs), warpPoint(wPoint), warpSeuil(wSeuil), noisePoint(nPoint), noiseSeuil(nSeuil)
+Octave::Octave(double f, double a, double ms, double as, double fs, FoncteurWarp wPoint, FoncteurWarp wSeuil, Noise nPoint, Noise nSeuil):freq(f), ampli(a), minSeuil(ms), ampliS(as), freqS(fs), warpPoint(wPoint), warpSeuil(wSeuil), noisePoint(nPoint), noiseSeuil(nSeuil)
 {
 }
 
-float Octave::getHauteur(const Vector2D &p)
+double Octave::getHauteur(const Vector2D &p)
 {
     Vector2D pSurFreq = p/freq;
     Vector2D pFreqS = p/freqS;
     Vector2D wPoint = warpPoint(pSurFreq);
     Vector2D wSeuil = warpSeuil(pFreqS);
-    float r = MathUtils::ridge(ampli*noisePoint.pointToValue(wPoint),
+    double r = MathUtils::ridge(ampli*noisePoint.pointToValue(wPoint),
                 minSeuil+ampliS*noiseSeuil.pointToValue(wSeuil));
     return r;
 }
